Add a couple redundant casts to prevent compile failures on certain platforms
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@247 37977431-3df6-0310-b722-df95706aa16b

+++ rts/BFGroundDrawer.cpp
@@ -939,7 +939,7 @@ bool CBFGroundDrawer::UpdateTextures()
 					float m=md->moveMath->SpeedMod(*md, x*2,y*2);
 					if(gs->cheatEnabled && md->moveMath->IsBlocked2(*md, x*2+1, y*2+1) & (CMoveMath::BLOCK_STRUCTURE | CMoveMath::BLOCK_TERRAIN))
 						m=0;
-					m=std::min(1.,sqrt(m));
+					m=std::min(1.,(double)sqrt(m));
 
 					infoTexMem[a*4+0]=255-int(m)*255;
 					infoTexMem[a*4+1]=int(m)*255;
@@ -951,7 +951,7 @@ bool CBFGroundDrawer::UpdateTextures()
 				for(int x=0;x<gs->hmapx;++x){
 					int a=y*(gs->pwr2mapx>>1)+x;
 					if(myAirLos[(y/2)*gs->hmapx/2+x/2])
-						infoTexMem[a*4]=(unsigned char)std::min(255.,sqrt(sqrt(extractDepthMap[y*gs->hmapx+x]))*900);
+						infoTexMem[a*4]=(unsigned char)std::min(255.,(double)sqrt(sqrt(extractDepthMap[y*gs->hmapx+x]))*900);
 					else
 						infoTexMem[a*4]=0;
 					infoTexMem[a*4+1]=(extraTexPal[extraTex[y*gs->hmapx+x]*3+1]);
